fix(spotify): move to pipelines
